reorganize collection of scripts
