+++ src/scripts/inferencing/lightgbm_c_api/score.py
@@ -18,7 +18,7 @@ from subprocess import TimeoutExpired
 
 # Add the right path to PYTHONPATH
 # so that you can import from common.*
-COMMON_ROOT = os.path.abspath(os.path.join(os.path.dirname(__file__), "..", ".."))
+COMMON_ROOT = os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "..", ".."))
 
 if COMMON_ROOT not in sys.path:
     logging.info(f"Adding {COMMON_ROOT} to PYTHONPATH")
