+++ src/scripts/inferencing/lightgbm_c_api/spec.additional_includes
@@ -1,0 +1,2 @@
+../../../common/
+../../../../docker/
