+++ src/scripts/inferencing/lightgbm_c_api/spec.yaml
