+++ src/scripts/inferencing/lightgbm_cli/spec.additional_includes
@@ -1,0 +1,2 @@
+../../../common/
+../../../../docker/
