+++ src/scripts/inferencing/lightgbm_cli/spec.yaml
