+++ src/scripts/inferencing/lightgbm_python/score.py
@@ -14,7 +14,7 @@ import numpy
 
 # Add the right path to PYTHONPATH
 # so that you can import from common.*
-COMMON_ROOT = os.path.abspath(os.path.join(os.path.dirname(__file__), "..", ".."))
+COMMON_ROOT = os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "..", ".."))
 
 if COMMON_ROOT not in sys.path:
     logging.info(f"Adding {COMMON_ROOT} to PYTHONPATH")
