+++ src/scripts/inferencing/lightgbm_python/spec.yaml
