+++ src/scripts/inferencing/treelite_python/spec.additional_includes
@@ -1,0 +1,1 @@
+../../../common/
