+++ src/scripts/inferencing/treelite_python/spec.yaml
