--- src/scripts/lightgbm_c_api/spec.additional_includes
@@ -1,2 +1,0 @@
-../../common/
-../../../docker/
