--- src/scripts/lightgbm_cli/score_spec.additional_includes
@@ -1,2 +1,0 @@
-../../common/
-../../../docker/
