--- src/scripts/lightgbm_python/score_spec.additional_includes
@@ -1,2 +1,0 @@
-../../common/
-../../../docker/
