--- src/scripts/lightgbm_python/train_spec.additional_includes
@@ -1,2 +1,0 @@
-../../common/
-../../../docker/
