+++ src/scripts/model_transformation/lightgbm_to_treelite/conda_env.yaml
@@ -1,0 +1,11 @@
+name: treelite_conda_env
+channels:
+- defaults
+dependencies:
+- python=3.8
+- pip=20.0
+- pip:
+  - azureml-defaults==1.30.0
+  - azureml-mlflow==1.30.0
+  - treelite==1.3.0
+  - treelite_runtime==1.3.0
