+++ src/scripts/training/lightgbm_python/spec.additional_includes
@@ -1,0 +1,2 @@
+../../../common/
+../../../../docker/
