+++ src/scripts/training/lightgbm_python/train.py
@@ -17,7 +17,7 @@ from collections import namedtuple
 
 # Add the right path to PYTHONPATH
 # so that you can import from common.*
-COMMON_ROOT = os.path.abspath(os.path.join(os.path.dirname(__file__), "..", ".."))
+COMMON_ROOT = os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "..", ".."))
 
 if COMMON_ROOT not in sys.path:
     print(f"Adding {COMMON_ROOT} to PYTHONPATH")
