--- src/scripts/treelite_python/score_spec.additional_includes
@@ -1,1 +1,0 @@
-../../common/
